Move create_async_openai_client into test_mcp_server.py (its only caller)
